services/mood-analysis-service/requirements.txt
grpcio==1.60.1
grpcio-tools==1.60.1
protobuf==4.25.1
confluent-kafka==2.3.0
torch==2.1.0
transformers==4.35.2
sentencepiece==0.1.99
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
pydantic==2.4.2
loguru==0.7.2